# .api/apis/chainabuse/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "basic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Chainabuse API",
    "version": "0.1.0",
    "description": "<p>The API uses Basic Authentication to authenticate requests. You can generate a new API key from your user profile on <a href='https://www.chainabuse.com'>Chainabuse</a>.</p><p>Once you have an API key, you have three options to authenticate:</p><ul><li>Click the green Authorize button at the top of the docs, and add your API key as the username (you can leave password blank). In this case, you will be authenticated against all API calls in the docs during your session.</li><li>Attempt an API call from your browser window and have your browser pop up a native Username/Password window. API key again goes in the username field (you can leave password blank).</li><li>Using curl, you can supply the api key as username (you can leave password blank) for basic auth. For example, if your api key is <code>ca_YlY1SnZsRWg4S0U0R0p5NG1TekR0Y251LjE2Tjc2OGFENnI2T0hhNzFYalFNUEE9PQ</code>, then you can run <code>curl https://api.chainabuse.com/v0/reports  --user ca_YlY1SnZsRWg4S0U0R0p5NG1TekR0Y251LjE2Tjc2OGFENnI2T0hhNzFYalFNUEE9PQ:</code> to get the list of reports.</li></ul>",
    "contact": {
      "email": "hello@chainabuse.com"
    }
  },
  "openapi": "3.0.0",
  "paths": {
    "/reports": {
      "get": {
        "operationId": "Reports",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "evidences": {
                        "items": {
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "ips": {
                        "items": {
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "losses": {
                        "items": {
                          "properties": {
                            "amount": {
                              "type": "number",
                              "format": "double"
                            },
                            "asset": {
                              "type": "string",
                              "enum": [
                                "OTHER",
                                "BTC",
                                "ETH",
                                "TRON",
                                "SOL",
                                "POLYGON",
                                "BINANCE",
                                "HBAR",
                                "CARDANO",
                                "LITECOIN",
                                "AVALANCHE",
                                "MULTIVERSX",
                                "ARBITRUM",
                                "ALGORAND",
                                "BASE",
                                "TON",
                                "AED",
                                "AFN",
                                "ALL",
                                "AMD",
                                "ANG",
                                "AOA",
                                "ARS",
                                "AUD",
                                "AWG",
                                "AZN",
                                "BAM",
                                "BBD",
                                "BDT",
                                "BGN",
                                "BHD",
                                "BIF",
                                "BMD",
                                "BND",
                                "BOB",
                                "BOV",
                                "BRL",
                                "BSD",
                                "BTN",
                                "BWP",
                                "BYN",
                                "BZD",
                                "CAD",
                                "CDF",
                                "CHE",
                                "CHF",
                                "CHW",
                                "CLF",
                                "CLP",
                                "CNY",
                                "COP",
                                "COU",
                                "CRC",
                                "CUC",
                                "CUP",
                                "CVE",
                                "CZK",
                                "DJF",
                                "DKK",
                                "DOP",
                                "DZD",
                                "EGP",
                                "ERN",
                                "ETB",
                                "EUR",
                                "FJD",
                                "FKP",
                                "GBP",
                                "GEL",
                                "GHS",
                                "GIP",
                                "GMD",
                                "GNF",
                                "GTQ",
                                "GYD",
                                "HKD",
                                "HNL",
                                "HRK",
                                "HTG",
                                "HUF",
                                "IDR",
                                "ILS",
                                "INR",
                                "IQD",
                                "IRR",
                                "ISK",
                                "JMD",
                                "JOD",
                                "JPY",
                                "KES",
                                "KGS",
                                "KHR",
                                "KMF",
                                "KPW",
                                "KRW",
                                "KWD",
                                "KYD",
                                "KZT",
                                "LAK",
                                "LBP",
                                "LKR",
                                "LRD",
                                "LSL",
                                "LYD",
                                "MAD",
                                "MDL",
                                "MGA",
                                "MKD",
                                "MMK",
                                "MNT",
                                "MOP",
                                "MRU",
                                "MUR",
                                "MVR",
                                "MWK",
                                "MXN",
                                "MXV",
                                "MYR",
                                "MZN",
                                "NAD",
                                "NGN",
                                "NIO",
                                "NOK",
                                "NPR",
                                "NZD",
                                "OMR",
                                "PAB",
                                "PEN",
                                "PGK",
                                "PHP",
                                "PKR",
                                "PLN",
                                "PYG",
                                "QAR",
                                "RON",
                                "RSD",
                                "RUB",
                                "RWF",
                                "SAR",
                                "SBD",
                                "SCR",
                                "SDG",
                                "SEK",
                                "SGD",
                                "SHP",
                                "SLL",
                                "SOS",
                                "SRD",
                                "SSP",
                                "STN",
                                "SVC",
                                "SYP",
                                "SZL",
                                "THB",
                                "TJS",
                                "TMT",
                                "TND",
                                "TOP",
                                "TRY",
                                "TTD",
                                "TWD",
                                "TZS",
                                "UAH",
                                "UGX",
                                "USD",
                                "USN",
                                "UYI",
                                "UYU",
                                "UYW",
                                "UZS",
                                "VES",
                                "VND",
                                "VUV",
                                "WST",
                                "XAF",
                                "XAG",
                                "XAU",
                                "XBA",
                                "XBB",
                                "XBC",
                                "XBD",
                                "XCD",
                                "XDR",
                                "XOF",
                                "XPD",
                                "XPF",
                                "XPT",
                                "XSU",
                                "XTS",
                                "XUA",
                                "XXX",
                                "YER",
                                "ZAR",
                                "ZMW",
                                "ZWL"
                              ],
                              "nullable": false
                            }
                          },
                          "required": [
                            "amount",
                            "asset"
                          ],
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "accusedScammers": {
                        "items": {
                          "properties": {
                            "info": {
                              "items": {
                                "properties": {
                                  "type": {
                                    "enum": [
                                      "NAME",
                                      "EMAIL",
                                      "PHONE",
                                      "DISCORD",
                                      "TELEGRAM",
                                      "TWITTER",
                                      "INSTAGRAM",
                                      "GITHUB",
                                      "FACEBOOK",
                                      "WHATSAPP",
                                      "TIKTOK",
                                      "YOUTUBE",
                                      "OTHER"
                                    ],
                                    "type": "string"
                                  },
                                  "contact": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "type",
                                  "contact"
                                ],
                                "type": "object"
                              },
                              "type": "array"
                            }
                          },
                          "required": [
                            "info"
                          ],
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "descriptionMarkdown": {
                        "type": "string"
                      },
                      "addresses": {
                        "items": {
                          "properties": {
                            "domain": {
                              "type": "string",
                              "nullable": true
                            },
                            "chain": {
                              "type": "string",
                              "nullable": true
                            },
                            "address": {
                              "type": "string",
                              "nullable": true
                            }
                          },
                          "required": [
                            "domain",
                            "address"
                          ],
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "description": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "scamCategory": {
                        "type": "string",
                        "nullable": true
                      },
                      "isPrivate": {
                        "type": "boolean"
                      },
                      "checked": {
                        "type": "boolean"
                      },
                      "trusted": {
                        "type": "boolean",
                        "description": "Reported by a contributor registered as trusted on Chainabuse."
                      },
                      "id": {
                        "type": "string",
                        "description": "Report ID"
                      }
                    },
                    "required": [
                      "addresses",
                      "createdAt",
                      "trusted",
                      "id"
                    ],
                    "type": "object",
                    "example": {
                      "id": "52907745-7672-470e-a803-a2f8feb52944",
                      "trusted": true,
                      "checked": true,
                      "scamCategory": "RUG_PULL",
                      "createdAt": "2022-09-09T04:53:16.591Z",
                      "addresses": [
                        {
                          "address": "12QeMLzSrB8XH8FvEzPMVoRxVAzTr5XM2y",
                          "chain": "BTC",
                          "domain": null
                        },
                        {
                          "address": null,
                          "chain": null,
                          "domain": "scammer.com"
                        }
                      ]
                    }
                  },
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "reason"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "reason"
                  ],
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Get Reports",
        "security": [
          {
            "basic": []
          }
        ],
        "parameters": [
          {
            "description": "(Optional if a domain is passed). Crypto addresses to screen.",
            "in": "query",
            "name": "address",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional. Threshold date to use as a filter.",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Optional. Scam category to use as filter.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "enum": [
                "RUG_PULL",
                "UKRANIAN_DONATION_SCAM",
                "DONATION_SCAM",
                "SEXTORTION",
                "SIM_SWAP",
                "PHISHING",
                "RANSOMWARE",
                "CONTRACT_EXPLOIT",
                "AIRDROP",
                "ROMANCE",
                "PIGBUTCHERING",
                "OTHER",
                "FAKE_PROJECT",
                "OTHER_BLACKMAIL",
                "IMPERSONATION",
                "FAKE_RETURNS",
                "UPGRADE_SCAM",
                "MAN_IN_THE_MIDDLE_ATTACK",
                "OTHER_HACK",
                "OTHER_INVESTMENT_SCAM"
              ],
              "type": "string"
            }
          },
          {
            "description": "(Optional if a URL is passed). Chain to use as filter.",
            "in": "query",
            "name": "chain",
            "required": false,
            "schema": {
              "enum": [
                "BTC",
                "ETH",
                "TRON",
                "SOL",
                "POLYGON",
                "BINANCE",
                "HBAR",
                "CARDANO",
                "LITECOIN",
                "AVALANCHE",
                "MULTIVERSX",
                "ARBITRUM",
                "ALGORAND",
                "BASE",
                "TON"
              ],
              "type": "string"
            }
          },
          {
            "description": "Optional\n- True: report checked by our team of moderators including blockchain intelligence experts.\n- False: report could not be verified by our team of moderators.",
            "in": "query",
            "name": "checked",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Optional. An ISO 3166-1 alpha-2 code e.g. AE for United Arab Emirates, CA for Canada, GB for United Kingdom, etc.\n- Usage: ?country=CA to filter reports within Canada",
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "(Optional if an address is passed). URL to screen.",
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional. Defaults to false. Only users with TRM_INTERNAL role can toggle.\n- True: private reports are included in the response.\n- False: private reports are hidden from the response.",
            "in": "query",
            "name": "includePrivate",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "description": "Optional. Filters reports with at least the amount passed (Currently, only usd asset are supported and others will be ignored).",
            "in": "query",
            "name": "minLossAmount",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number",
              "minimum": 0
            }
          },
          {
            "description": "- ASC = latest reports first\n- DESC= oldest report first if several reports are pulled out",
            "in": "query",
            "name": "orderByDirection",
            "required": false,
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "description": "Orders reports by date if several reports are pulled out",
            "in": "query",
            "name": "orderByField",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "CREATED_AT",
                "SUBMITTED_BY"
              ]
            }
          },
          {
            "description": "Optional. Number of pages of reports to pull out. The maximum number of reports per page is 50. If page =1, you will pull out a maximum of 50 reports.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "double",
              "type": "number",
              "minimum": 1
            }
          },
          {
            "description": "Optional. Number of reports to display per page. The maximum number of reports per page is 50. If page =1, you will pull out a maximum of 50 reports.",
            "in": "query",
            "name": "perPage",
            "required": false,
            "schema": {
              "default": 50,
              "format": "double",
              "type": "number",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "description": "Optional. Filters reports based on passed indicator of compromise.",
            "in": "query",
            "name": "scammerIoc",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional. Start date to use as a filter.",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Optional\n- True: reported by a trusted contributor, based on how this Partner detects and verifies information.\n- False: the reporter is not registered as a trusted contributor. Please note this does not mean their report cannot be trusted.",
            "in": "query",
            "name": "trusted",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Optional. The contributor's username.",
            "in": "query",
            "name": "username",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.chainabuse.com/v0"
    }
  ]
}